Move deletion confirmation helpers to pkg/utils/gardener
