+++ pkg/extensions/customresources.go
@@ -27,9 +27,9 @@ import (
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	gardencorev1beta1helper "github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
-	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/utils"
 	"github.com/gardener/gardener/pkg/utils/flow"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/kubernetes/health"
 	"github.com/gardener/gardener/pkg/utils/retry"
@@ -148,7 +148,7 @@ func DeleteExtensionCR(
 	obj.SetNamespace(namespace)
 	obj.SetName(name)
 
-	if err := common.ConfirmDeletion(ctx, c, obj); err != nil {
+	if err := gutil.ConfirmDeletion(ctx, c, obj); err != nil {
 		return err
 	}
