+++ pkg/operation/botanist/component/extensions/extension/extension_test.go
@@ -28,7 +28,7 @@ import (
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	mocktime "github.com/gardener/gardener/pkg/mock/go/time"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/extensions/extension"
-	"github.com/gardener/gardener/pkg/operation/common"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/test"
 	. "github.com/gardener/gardener/pkg/utils/test/matchers"
@@ -179,7 +179,10 @@ var _ = Describe("Extension", func() {
 		})
 
 		It("should return error if not deleted successfully", func() {
-			defer test.WithVars(&extensions.TimeNow, mockNow.Do)()
+			defer test.WithVars(
+				&extensions.TimeNow, mockNow.Do,
+				&gutil.TimeNow, mockNow.Do,
+			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
 			expectedExtension := extensionsv1alpha1.Extension{
@@ -187,7 +190,7 @@ var _ = Describe("Extension", func() {
 					Name:      "ext1",
 					Namespace: namespace,
 					Annotations: map[string]string{
-						common.ConfirmationDeletion:        "true",
+						gutil.ConfirmationDeletion:         "true",
 						v1beta1constants.GardenerTimestamp: now.UTC().String(),
 					},
 				},
