+++ pkg/operation/botanist/component/extensions/infrastructure/infrastructure_test.go
@@ -28,7 +28,7 @@ import (
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	mocktime "github.com/gardener/gardener/pkg/mock/go/time"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/extensions/infrastructure"
-	"github.com/gardener/gardener/pkg/operation/common"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/retry"
 	retryfake "github.com/gardener/gardener/pkg/utils/retry/fake"
@@ -294,12 +294,13 @@ var _ = Describe("#Interface", func() {
 		It("should not return error when it's deleted successfully", func() {
 			defer test.WithVars(
 				&extensions.TimeNow, mockNow.Do,
+				&gutil.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
 			infraCopy := infra.DeepCopy()
 			infraCopy.Annotations = map[string]string{
-				common.ConfirmationDeletion:        "true",
+				gutil.ConfirmationDeletion:         "true",
 				v1beta1constants.GardenerTimestamp: now.UTC().String(),
 			}
 
@@ -319,12 +320,13 @@ var _ = Describe("#Interface", func() {
 		It("should return error when it's not deleted successfully", func() {
 			defer test.WithVars(
 				&extensions.TimeNow, mockNow.Do,
+				&gutil.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
 			infraCopy := infra.DeepCopy()
 			infraCopy.Annotations = map[string]string{
-				common.ConfirmationDeletion:        "true",
+				gutil.ConfirmationDeletion:         "true",
 				v1beta1constants.GardenerTimestamp: now.UTC().String(),
 			}
