+++ pkg/operation/common/utils_test.go
@@ -16,17 +16,13 @@ package common_test
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net"
 	"strings"
-	"time"
 
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
-	mocktime "github.com/gardener/gardener/pkg/mock/go/time"
 	. "github.com/gardener/gardener/pkg/operation/common"
-	"github.com/gardener/gardener/pkg/utils/test"
 
 	hvpav1alpha1 "github.com/gardener/hvpa-controller/api/v1alpha1"
 	"github.com/golang/mock/gomock"
@@ -218,144 +214,6 @@ var _ = Describe("common", func() {
 		Entry("with special characters", "foo", `C*ko4P++$"x`, `"$++*ab*$c4k`),
 		Entry("with special characters", "foo", "P+*4", `P*$8uOkv6+4`),
 	)
-
-	Describe("#CheckIfDeletionIsConfirmed", func() {
-		It("should prevent the deletion due to missing annotations", func() {
-			obj := &corev1.Namespace{}
-
-			Expect(CheckIfDeletionIsConfirmed(obj)).To(HaveOccurred())
-		})
-
-		It("should prevent the deletion due annotation value != true", func() {
-			obj := &corev1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						ConfirmationDeletion: "false",
-					},
-				},
-			}
-
-			Expect(CheckIfDeletionIsConfirmed(obj)).To(HaveOccurred())
-		})
-
-		It("should allow the deletion due annotation value == true", func() {
-			obj := &corev1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						ConfirmationDeletion: "true",
-					},
-				},
-			}
-
-			Expect(CheckIfDeletionIsConfirmed(obj)).To(Succeed())
-		})
-	})
-
-	Describe("#ConfirmDeletion", func() {
-		var (
-			ctrl    *gomock.Controller
-			c       *mockclient.MockClient
-			now     time.Time
-			mockNow *mocktime.MockNow
-		)
-
-		BeforeEach(func() {
-			ctrl = gomock.NewController(GinkgoT())
-			mockNow = mocktime.NewMockNow(ctrl)
-			c = mockclient.NewMockClient(ctrl)
-		})
-
-		AfterEach(func() {
-			ctrl.Finish()
-		})
-
-		It("should add the deletion confirmation annotation for an object without annotations", func() {
-			var (
-				ctx = context.TODO()
-				obj = &corev1.Namespace{}
-			)
-
-			defer test.WithVars(
-				&TimeNow, mockNow.Do,
-			)()
-
-			expectedObj := obj.DeepCopy()
-			expectedObj.Annotations = map[string]string{ConfirmationDeletion: "true", v1beta1constants.GardenerTimestamp: now.UTC().String()}
-
-			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
-
-			c.EXPECT().Get(ctx, gomock.AssignableToTypeOf(client.ObjectKey{}), obj)
-			c.EXPECT().Update(ctx, expectedObj)
-
-			Expect(ConfirmDeletion(ctx, c, obj)).To(Succeed())
-		})
-
-		It("should add the deletion confirmation annotation for an object with annotations", func() {
-			var (
-				ctx = context.TODO()
-				obj = &corev1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Annotations: map[string]string{
-							"foo": "bar",
-						},
-					},
-				}
-			)
-
-			defer test.WithVars(
-				&TimeNow, mockNow.Do,
-			)()
-
-			expectedObj := obj.DeepCopy()
-			expectedObj.Annotations[ConfirmationDeletion] = "true"
-			expectedObj.Annotations[v1beta1constants.GardenerTimestamp] = now.UTC().String()
-
-			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
-			c.EXPECT().Get(ctx, gomock.AssignableToTypeOf(client.ObjectKey{}), obj)
-			c.EXPECT().Update(ctx, expectedObj)
-
-			Expect(ConfirmDeletion(ctx, c, obj)).To(Succeed())
-		})
-
-		It("should ignore non-existing objects", func() {
-			var (
-				ctx         = context.TODO()
-				obj         = &corev1.Namespace{}
-				expectedObj = obj.DeepCopy()
-			)
-
-			c.EXPECT().Get(ctx, gomock.AssignableToTypeOf(client.ObjectKey{}), obj).Return(apierrors.NewNotFound(corev1.Resource("namespaces"), ""))
-
-			Expect(ConfirmDeletion(ctx, c, obj)).To(Succeed())
-			Expect(obj).To(Equal(expectedObj))
-		})
-
-		It("should retry on conflict and add the deletion confirmation annotation", func() {
-			var (
-				ctx     = context.TODO()
-				baseObj = &corev1.Namespace{}
-				obj     = baseObj.DeepCopy()
-			)
-
-			defer test.WithVars(
-				&TimeNow, mockNow.Do,
-			)()
-
-			expectedObj := obj.DeepCopy()
-			expectedObj.Annotations = map[string]string{ConfirmationDeletion: "true", v1beta1constants.GardenerTimestamp: now.UTC().String()}
-
-			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
-			c.EXPECT().Get(ctx, gomock.AssignableToTypeOf(client.ObjectKey{}), obj)
-			c.EXPECT().Update(ctx, expectedObj).Return(apierrors.NewConflict(corev1.Resource("namespaces"), "", errors.New("conflict")))
-			c.EXPECT().Get(ctx, gomock.AssignableToTypeOf(client.ObjectKey{}), expectedObj).DoAndReturn(func(_ context.Context, _ client.ObjectKey, obj client.Object) error {
-				baseObj.DeepCopyInto(obj.(*corev1.Namespace))
-				return nil
-			})
-			c.EXPECT().Update(ctx, expectedObj)
-
-			Expect(ConfirmDeletion(ctx, c, obj)).To(Succeed())
-		})
-	})
 
 	Describe("#ExtensionID", func() {
 		It("should return the expected identifier", func() {
